Prepeare for merge with DALA-7363 (too many functions in one file)
